+++ events/faq_ready.js
@@ -1,10 +1,11 @@
 // @ts-check
 const { DBFaq } = require("../lib/database/faq");
+/* eslint-disable */
 
 module.exports = {
     name: "ready",
     once: true,
-    async execute(client) {
+    async execute() {
         const faqStorage = new DBFaq();
         global.faqStorage = faqStorage;
     },
