+++ events/reactrole_read.js
@@ -1,9 +1,10 @@
 const { DBReactRole } = require("../lib/database/dbreactrole");
+/* eslint-disable */
 
 module.exports = {
     name: "ready",
     once: true,
-    async execute(client) {
+    async execute() {
         const reactRoles = new DBReactRole();
         global.reactRoles = reactRoles;
     },
