+++ events/schedulepost_ready.js
@@ -1,9 +1,10 @@
 const { DBSchedulePost } = require("../lib/database/dbschedulepost");
+/* eslint-disable */
 
 module.exports = {
     name: "ready",
     once: true,
-    async execute(client) {
+    async execute() {
         const schedulePost = new DBSchedulePost();
         global.schedulePost = schedulePost;
     },
