+++ lib/carrotboard/index.js
@@ -1,14 +1,8 @@
 // @ts-check
-
+/* eslint-disable */
 // const { Embed } = require("@discordjs/builders");
 const {
-    Client,
-    MessageReaction,
     MessageEmbed,
-    MessageEmbedOptions,
-    Channel,
-    User,
-    Message,
 } = require("discord.js");
 const { DBcarrotboard } = require("../database/dbcarrotboard");
 const fs = require("fs");
@@ -194,7 +188,8 @@ class CarrotboardStorage {
                     inline: true,
                 },
                 {
-                    name: "Message" + "\u2800".repeat(6), // 20 for full but messed up on thin display
+                    // 20 for full but messed up on thin display
+                    name: "Message" + "\u2800".repeat(6),
                     value: `${content}\n\u200b`,
                     inline: true,
                 },
