+++ lib/connect4/connect4Runner.js
@@ -1,5 +1,7 @@
 const { connect4GameObj } = require("./connect4Game");
 
+/* eslint-disable */
+
 const maxNumGames = 1000;
 const symbols = ["⬛", "🔴", "🟡"];
 const keycapEmojis = {
