+++ lib/database/database.js
@@ -226,13 +226,13 @@ class DBuser {
             const result = await client.query(query, values);
 
             if (result.rows.length != 0) {
-                let query = "UPDATE users SET joindate=$1, userleft=$2 where userid=$3";
+                query = "UPDATE users SET joindate=$1, userleft=$2 where userid=$3";
                 values = [time, false, userid];
                 await client.query(query, values);
             } else {
-                let query =
+                query =
                     "INSERT INTO users (USERID, JOINDATE, LEAVEDATE, USERLEFT) VALUES ($1,$2,$3,$4)";
-                let values = [userid, time, null, false];
+                values = [userid, time, null, false];
                 await client.query(query, values);
             }
             await client.query("COMMIT");
@@ -277,7 +277,7 @@ class DBuser {
             const values = [userid, role];
             const query = `DELETE FROM user_roles
         where userid = $1 and role = $2`;
-            const result = await client.query(query, values);
+            await client.query(query, values);
             await client.query("COMMIT");
         } catch (ex) {
             console.log(`Something wrong happend ${ex}`);
@@ -361,7 +361,7 @@ class DBuser {
             const values = [userid, permission];
             const query = `DELETE FROM user_permissions
         where userid = $1 and permission = $2 `;
-            const result = await client.query(query, values);
+            await client.query(query, values);
 
             // console.log(result.rows);
             await client.query("COMMIT");
@@ -386,7 +386,7 @@ class DBuser {
 
             query = "INSERT INTO user_channels (CID, USERID, CHANNEL) VALUES ($1,$2,$3)";
             const values = [count, userid, channel];
-            let result = await client.query(query, values);
+            result = await client.query(query, values);
 
             await client.query("COMMIT");
         } catch (ex) {
@@ -406,7 +406,7 @@ class DBuser {
             const values = [userid, channel];
             const query = `DELETE FROM user_channels
         where userid = $1 and channel = $2 `;
-            const result = await client.query(query, values);
+            await client.query(query, values);
 
             await client.query("COMMIT");
         } catch (ex) {
@@ -458,10 +458,7 @@ class DBuser {
     }
 }
 
-// For testing
-/*
-(async () => {
-    let as = new DBuser();
-})();
+module.exports = {
+    DBuser,
+};
 
-*/
