+++ lib/database/dbcarrotboard.js
@@ -82,7 +82,7 @@ class DBcarrotboard {
             MESSAGE_CONTENTS CHAR(50)
             )`;
 
-            const result = await client.query(query);
+            await client.query(query);
             await client.query("COMMIT");
         } catch (ex) {
             console.log(`Something wrong happend ${ex}`);
@@ -153,7 +153,7 @@ class DBcarrotboard {
                     "INSERT INTO carrot_board (EMOJI, MESSAGE_ID, USER_ID, CHANNEL_ID, COUNT, MESSAGE_CONTENTS) VALUES ($1,$2,$3,$4,$5,$6)";
                 const values = [emoji, message_id, user_id, channel_id, 1, message_contents];
 
-                const result = await client.query(query, values);
+                await client.query(query, values);
                 await client.query("COMMIT");
             } else {
                 let count = await this.get_count(emoji, message_id, user_id, channel_id);
@@ -162,7 +162,7 @@ class DBcarrotboard {
                     "UPDATE carrot_board SET count = $1  where emoji = $2 and message_id = $3 and user_id = $4 and channel_id = $5";
                 const values = [count, emoji, message_id, user_id, channel_id];
 
-                const result = await client.query(query, values);
+                await client.query(query, values);
                 await client.query("COMMIT");
             }
         } catch (ex) {
@@ -272,7 +272,7 @@ class DBcarrotboard {
             const query = "DELETE FROM carrot_board where message_id = $1 and channel_id = $2";
 
             const values = [message_id, channel_id];
-            const result = await client.query(query, values);
+            await client.query(query, values);
             await client.query("COMMIT");
         } catch (ex) {
             console.log(`Something wrong happend ${ex}`);
@@ -293,7 +293,7 @@ class DBcarrotboard {
                 "DELETE FROM carrot_board where message_id = $1 and channel_id = $2  and user_id = $3 and emoji = $4";
 
             const values = [message_id, channel_id, user_id, emoji];
-            const result = await client.query(query, values);
+            await client.query(query, values);
             await client.query("COMMIT");
         } catch (ex) {
             console.log(`Something wrong happend ${ex}`);
@@ -320,7 +320,7 @@ class DBcarrotboard {
                     "UPDATE carrot_board SET count = $1  where emoji = $2 and message_id = $3 and user_id = $4 and channel_id = $5";
 
                 const values = [count, emoji, message_id, user_id, channel_id];
-                const result = await client.query(query, values);
+                await client.query(query, values);
                 await client.query("COMMIT");
             }
         } catch (ex) {
