+++ lib/database/faq.js
@@ -50,7 +50,8 @@ class DBFaq {
             );
             await client.query("COMMIT");
 
-            return result.rowCount > 0; // return whether there was a table or not
+            // return whether there was a table or not
+            return result.rowCount > 0;
         } catch (err) {
             console.log(`Something went wrong ${err}`);
         } finally {
@@ -71,7 +72,7 @@ class DBFaq {
                 TAGS TEXT
             )`;
 
-            const result = await client.query(query);
+            await client.query(query);
             await client.query("COMMIT");
         } catch (err) {
             console.log(`Something went wrong ${err}`);
