+++ lib/discordscroll/scroller.js
@@ -1,12 +1,8 @@
 const {
     MessageEmbed,
-    Message,
     CommandInteraction,
-    InteractionReplyOptions,
     MessageButton,
     MessageActionRow,
-    InteractionCollector,
-    MessageComponentInteraction,
 } = require("discord.js");
 
 class DiscordScroll {
@@ -176,7 +172,7 @@ class DiscordScroll {
             this._scroll(buttonInt);
         });
 
-        buttonCollector.on("end", (collected) => {
+        buttonCollector.on("end", () => {
             this._deactivate();
         });
     }
